restore compare hashes test
